requirements.txt
flask==2.3.2
werkzeug==2.3.8
twilio==7.12.0
gspread==5.4.0
google-auth==2.9.0
google-api-python-client==2.52.0
schedule==1.1.0
requests==2.28.1